Split some classes and const definitions out.
