+++ src/Makefile.am
@@ -99,6 +99,7 @@ BITCOIN_CORE_H = \
   compat/endian.h \
   compat/sanity.h \
   compressor.h \
+  consensus/consensus.h \
   consensus/merkle.h \
   primitives/block.h \
   primitives/transaction.h \
