+++ src/amount.cpp
@@ -8,6 +8,8 @@
 
 #include "tinyformat.h"
 
+const std::string CURRENCY_UNIT = "PIV";
+
 CFeeRate::CFeeRate(const CAmount& nFeePaid, size_t nSize)
 {
     if (nSize > 0)
