+++ src/wallet/wallet.cpp
@@ -8,6 +8,7 @@
 #include "wallet/wallet.h"
 
 #include "zpiv/accumulators.h"
+#include "amount.h"
 #include "base58.h"
 #include "checkpoints.h"
 #include "coincontrol.h"
